+++ packages/perseus-core/src/parse-perseus-json/perseus-parsers/interactive-graph-widget.ts
@@ -20,6 +20,8 @@ import {discriminatedUnionOn} from "../general-purpose-parsers/discriminated-uni
 import {parsePerseusImageBackground} from "./perseus-image-background";
 import {parseWidget} from "./widget";
 
+// TODO(LEMS-4224): don't import from outside of the parser
+// eslint-disable-next-line import/no-restricted-paths
 import type {PerseusGraphTypeLinear} from "../../data-schema";
 
 // Used to represent 2-D points and ranges
@@ -35,6 +37,7 @@ const parsePerseusGraphTypeAngle = object({
     coords: optional(trio(pairOfNumbers, pairOfNumbers, pairOfNumbers)),
     startCoords: optional(trio(pairOfNumbers, pairOfNumbers, pairOfNumbers)),
     pointLabels: optional(trio(string, string, string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeCircle = object({
@@ -48,6 +51,7 @@ const parsePerseusGraphTypeCircle = object({
         }),
     ),
     pointLabels: optional(array(string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeLinear = object({
@@ -55,6 +59,7 @@ const parsePerseusGraphTypeLinear = object({
     coords: optional(nullable(pair(pairOfNumbers, pairOfNumbers))),
     startCoords: optional(pair(pairOfNumbers, pairOfNumbers)),
     pointLabels: optional(pair(string, string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeLinearSystem = object({
@@ -63,6 +68,7 @@ const parsePerseusGraphTypeLinearSystem = object({
     coords: optional(nullable(array(pair(pairOfNumbers, pairOfNumbers)))),
     startCoords: optional(array(pair(pairOfNumbers, pairOfNumbers))),
     pointLabels: optional(array(string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeNone = object({
@@ -76,6 +82,7 @@ const parsePerseusGraphTypePoint = object({
     startCoords: optional(array(pairOfNumbers)),
     coord: optional(pairOfNumbers),
     pointLabels: optional(array(string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypePolygon = object({
@@ -88,6 +95,7 @@ const parsePerseusGraphTypePolygon = object({
     startCoords: optional(array(pairOfNumbers)),
     coords: optional(nullable(array(pairOfNumbers))),
     pointLabels: optional(array(string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeQuadratic = object({
@@ -97,6 +105,7 @@ const parsePerseusGraphTypeQuadratic = object({
     ),
     startCoords: optional(trio(pairOfNumbers, pairOfNumbers, pairOfNumbers)),
     pointLabels: optional(trio(string, string, string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeRay = object({
@@ -104,6 +113,7 @@ const parsePerseusGraphTypeRay = object({
     coords: optional(nullable(pair(pairOfNumbers, pairOfNumbers))),
     startCoords: optional(pair(pairOfNumbers, pairOfNumbers)),
     pointLabels: optional(pair(string, string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeSegment = object({
@@ -113,6 +123,7 @@ const parsePerseusGraphTypeSegment = object({
     coords: optional(nullable(array(pair(pairOfNumbers, pairOfNumbers)))),
     startCoords: optional(array(pair(pairOfNumbers, pairOfNumbers))),
     pointLabels: optional(array(string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeSinusoid = object({
@@ -120,6 +131,7 @@ const parsePerseusGraphTypeSinusoid = object({
     coords: optional(nullable(array(pairOfNumbers))),
     startCoords: optional(array(pairOfNumbers)),
     pointLabels: optional(array(string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeExponential = object({
@@ -133,6 +145,7 @@ const parsePerseusGraphTypeExponential = object({
         }),
     ),
     pointLabels: optional(array(string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeAbsoluteValue = object({
@@ -140,6 +153,7 @@ const parsePerseusGraphTypeAbsoluteValue = object({
     coords: optional(nullable(pair(pairOfNumbers, pairOfNumbers))),
     startCoords: optional(pair(pairOfNumbers, pairOfNumbers)),
     pointLabels: optional(pair(string, string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeTangent = object({
@@ -147,6 +161,7 @@ const parsePerseusGraphTypeTangent = object({
     coords: optional(nullable(array(pairOfNumbers))),
     startCoords: optional(array(pairOfNumbers)),
     pointLabels: optional(array(string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeLogarithm = object({
@@ -160,6 +175,7 @@ const parsePerseusGraphTypeLogarithm = object({
         }),
     ),
     pointLabels: optional(array(string)),
+    showPointLabels: optional(boolean),
 });
 
 const parsePerseusGraphTypeVector = object({